Implemented player in correct file
moved player implementation to player.cpp from source.cpp
